create separate grpc module, move organisation logic to lib from bin
