+++ organisation/src/core/mod.rs
@@ -1,3 +1,5 @@
 pub mod ethereum;
+pub mod grpc;
 pub mod ipfs;
+pub mod peer;
 pub mod protocol;
